refactor(titlebar): extract titlebar code into app/features/titlebar/
Move all titlebar logic (CSS, renderer script, icon helper, setup)
from main.js into a dedicated module for maintainability and to
prepare for macOS titlebar support.

Also fix latent bug: getElementById on element in disconnected DOM
tree (before prepend) — changed to querySelector on parent node.
