+++ app/features/titlebar/styles.js
@@ -1,0 +1,40 @@
+/**
+ * CSS for the custom in-page title bar (Windows).
+ */
+const TITLEBAR_CSS = ''
+    + '#sonacove-titlebar{position:fixed;top:0;left:0;right:0;height:32px;background:#1A1A1A;'
+    + 'border-bottom:2px solid;border-image:linear-gradient(90deg,#E8613C,#F59E0B) 1;'
+    + '-webkit-app-region:drag;display:flex;align-items:center;padding:0 12px;z-index:2147483647;'
+    + 'font-family:-apple-system,BlinkMacSystemFont,"Segoe UI",Roboto,sans-serif;font-size:12px;'
+    + 'color:#c0c0c0;user-select:none;box-sizing:border-box;}'
+    + '#sonacove-titlebar .stb-icon{width:20px;height:20px;margin-right:8px;background-size:contain;'
+    + 'background-repeat:no-repeat;background-position:center;}'
+    + '#sonacove-titlebar .stb-title{overflow:hidden;text-overflow:ellipsis;white-space:nowrap;}'
+    + '#sonacove-titlebar .stb-ver{font-size:10px;color:#555566;margin-left:4px;}'
+    + '#sonacove-titlebar .stb-ver.stb-update{color:#E8613C;font-weight:600;cursor:pointer;-webkit-app-region:no-drag;'
+    + 'background:#2A2A35;border:1px solid #E8613C;border-radius:10px;padding:2px 8px;font-size:10px;'
+    + 'transition:background 0.15s ease,color 0.15s ease;margin-left:8px;}'
+    + '#sonacove-titlebar .stb-ver.stb-update:hover{background:#3A2A2A;color:#F59E0B;border-color:#F59E0B;}'
+    + '#sonacove-titlebar .stb-update-dot{width:5px;height:5px;border-radius:50%;background:#E8613C;display:inline-block;margin-right:4px;vertical-align:middle;}'
+    + '#sonacove-titlebar .stb-menu{display:flex;gap:2px;-webkit-app-region:no-drag;}'
+    + '#sonacove-titlebar .stb-btn{background:transparent;border:none;color:#a0a0a0;cursor:pointer;'
+    + 'padding:4px 10px;border-radius:4px;font-size:12px;font-family:inherit;line-height:1;'
+    + 'transition:background 0.15s ease,color 0.15s ease;}'
+    + '#sonacove-titlebar .stb-btn:hover{background:rgba(255,255,255,0.1);color:#fff;}'
+    + '#sonacove-titlebar .stb-btn:active{background:rgba(255,255,255,0.18);color:#fff;}'
+    + '#sonacove-titlebar .stb-wc{display:flex;-webkit-app-region:no-drag;margin-left:8px;}'
+    + '#sonacove-titlebar .stb-wc-btn{background:transparent;border:none;color:#9090A0;cursor:pointer;'
+    + 'width:46px;height:30px;display:flex;align-items:center;justify-content:center;'
+    + '-webkit-app-region:no-drag;transition:background 0.15s ease,color 0.15s ease;}'
+    + '#sonacove-titlebar .stb-wc-btn svg{pointer-events:none;}'
+    + '#sonacove-titlebar .stb-wc-btn:hover{background:rgba(255,255,255,0.08);color:#D0D0DA;}'
+    + '#sonacove-titlebar .stb-wc-btn:active{background:rgba(255,255,255,0.14);color:#E0E0E6;}'
+    + '#sonacove-titlebar .stb-wc-btn.stb-close:hover{background:#e81123;color:#fff;}'
+    + 'html{box-sizing:border-box!important;padding-top:34px!important;overflow:hidden!important;}'
+    + 'body{height:calc(100vh - 34px)!important;overflow-y:auto!important;}'
+    + 'body::-webkit-scrollbar{width:8px;}'
+    + 'body::-webkit-scrollbar-track{background:transparent;}'
+    + 'body::-webkit-scrollbar-thumb{background:rgba(255,255,255,0.2);border-radius:4px;}'
+    + 'body::-webkit-scrollbar-thumb:hover{background:rgba(255,255,255,0.3);}';
+
+module.exports = { TITLEBAR_CSS };
